backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
dnspython==2.6.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
